+++ pkg/runtime/runtime.go
@@ -2,8 +2,9 @@ package runtime
 
 import (
 	"fmt"
-	"github.com/SimonBaeumer/commander/pkg/cmd"
+	"github.com/SimonBaeumer/cmd"
 	"log"
+	"os"
 	"runtime"
 	"strings"
 	"sync"
@@ -40,11 +41,12 @@ type TestCase struct {
 
 //TestConfig represents the configuration for a test
 type TestConfig struct {
-	Env      map[string]string
-	Dir      string
-	Timeout  string
-	Retries  int
-	Interval string
+	Env        map[string]string
+	Dir        string
+	Timeout    string
+	Retries    int
+	Interval   string
+	InheritEnv bool
 }
 
 // ResultStatus represents the status code of a test result
@@ -81,12 +83,13 @@ type ExpectedOut struct {
 
 // CommandUnderTest represents the command under test
 type CommandUnderTest struct {
-	Cmd      string
-	Env      map[string]string
-	Dir      string
-	Timeout  string
-	Retries  int
-	Interval string
+	Cmd        string
+	InheritEnv bool
+	Env        map[string]string
+	Dir        string
+	Timeout    string
+	Retries    int
+	Interval   string
 }
 
 // TestResult represents the TestCase and the ValidationResult
@@ -157,13 +160,22 @@ func executeRetryInterval(t TestCase) {
 
 // runTest executes the current test case
 func runTest(test TestCase) TestResult {
-	// cut = command under test
-	cut := cmd.NewCommand(test.Command.Cmd)
-	cut.SetTimeout(test.Command.Timeout)
-	cut.Dir = test.Command.Dir
-	for k, v := range test.Command.Env {
-		cut.AddEnv(k, v)
+	timeoutOpt, err := createTimeoutOption(test.Command.Timeout)
+	if err != nil {
+		test.Result = CommandResult{Error: err}
+		return TestResult{
+			TestCase: test,
+		}
 	}
+
+	envOpt := createEnvVarsOption(test)
+
+	// cut = command under test
+	cut := cmd.NewCommand(
+		test.Command.Cmd,
+		cmd.WithWorkingDir(test.Command.Dir),
+		timeoutOpt,
+		envOpt)
 
 	if err := cut.Execute(); err != nil {
 		log.Println(test.Title, " failed ", err.Error())
@@ -176,15 +188,15 @@ func runTest(test TestCase) TestResult {
 		}
 	}
 
-	log.Println("title: '"+test.Title+"'", " Command: ", cut.Cmd)
+	log.Println("title: '"+test.Title+"'", " Command: ", test.Command.Cmd)
 	log.Println("title: '"+test.Title+"'", " Directory: ", cut.Dir)
 	log.Println("title: '"+test.Title+"'", " Env: ", cut.Env)
 
 	// Write test result
 	test.Result = CommandResult{
 		ExitCode: cut.ExitCode(),
-		Stdout:   strings.Replace(cut.Stdout(), "\r\n", "\n", -1),
-		Stderr:   strings.Replace(cut.Stderr(), "\r\n", "\n", -1),
+		Stdout:   strings.TrimSpace(strings.Replace(cut.Stdout(), "\r\n", "\n", -1)),
+		Stderr:   strings.TrimSpace(strings.Replace(cut.Stderr(), "\r\n", "\n", -1)),
 	}
 
 	log.Println("title: '"+test.Title+"'", " ExitCode: ", test.Result.ExitCode)
@@ -192,6 +204,36 @@ func runTest(test TestCase) TestResult {
 	log.Println("title: '"+test.Title+"'", " Stderr: ", test.Result.Stderr)
 
 	return Validate(test)
+}
+
+func createEnvVarsOption(test TestCase) func(c *cmd.Command) {
+	return func(c *cmd.Command) {
+		// Add all env variables from parent process
+		if test.Command.InheritEnv {
+			for _, v := range os.Environ() {
+				split := strings.Split(v, "=")
+				c.AddEnv(split[0], split[1])
+			}
+		}
+
+		// Add custom env variables
+		for k, v := range test.Command.Env {
+			c.AddEnv(k, v)
+		}
+	}
+}
+
+func createTimeoutOption(timeout string) (func(c *cmd.Command), error) {
+	timeoutOpt := cmd.WithoutTimeout
+	if timeout != "" {
+		d, err := time.ParseDuration(timeout)
+		if err != nil {
+			return func(c *cmd.Command) {}, err
+		}
+		timeoutOpt = cmd.WithTimeout(d)
+	}
+
+	return timeoutOpt, nil
 }
 
 // GetRetries returns the retries of the command